+++ enc/fast_log.h
@@ -124,7 +124,7 @@ static const float kLog2Table[] = {
 #define LOG_2_INV 1.4426950408889634
 
 /* Faster logarithm for small integers, with the property of log2(0) == 0. */
-static BROTLI_INLINE double FastLog2(size_t v) {
+static BROTLI_INLINE float FastLog2(size_t v) {
   if (v < sizeof(kLog2Table) / sizeof(kLog2Table[0])) {
     return kLog2Table[v];
   }
@@ -132,9 +132,9 @@ static BROTLI_INLINE double FastLog2(size_t v) {
     (defined(__ANDROID_API__) && __ANDROID_API__ < 18)
   /* Visual Studio 2012 and Android API levels < 18 do not have the log2()
    * function defined, so we use log() and a multiplication instead. */
-  return log((double)v) * LOG_2_INV;
+  return log((float)v) * LOG_2_INV;
 #else
-  return log2((double)v);
+  return log2((float)v);
 #endif
 }
